wireless:libertas: fix race conditions on resume
on resume the card is powered up and the firmware is loaded
for initialisation.
If in that time something (like wpa_supplicant)
causes lbs_dev_open() to be called, it will fail
which irritates wpa_supplicant and the pm_runtime
refcounting will be wrong causing the card to not be
powered down on further ifconfig down calls.
The patch adds a mutex to prevent that.

Signed-off-by: Andreas Kemnade <andreas@kemnade.info>